third_party/WebKit/LayoutTests/inspector/tracing/timeline-js/timeline-script-id-expected.txt

Test that checks location resolving mechanics for TimerInstall TimerRemove and FunctionCall events with scriptId.

It expects two FunctionCall for InjectedScript, two TimerInstall events, two FunctionCall events and one TimerRemove event to be logged with performActions.js script name and some line number.

detailsTextContent for TimerInstall event: 'performActions.js:5'
details.textContent for TimerInstall event: 'performActions.js:5'
detailsTextContent for TimerInstall event: 'performActions.js:6'
details.textContent for TimerInstall event: 'performActions.js:6'
detailsTextContent for FunctionCall event: 'performActions.js:10'
details.textContent for FunctionCall event: 'intervalTimerWork @ performActions.js:10'
detailsTextContent for FunctionCall event: 'performActions.js:10'
details.textContent for FunctionCall event: 'intervalTimerWork @ performActions.js:10'
detailsTextContent for TimerRemove event: 'performActions.js:14'
details.textContent for TimerRemove event: 'performActions.js:14'
